+++ src/controllers/User.js
@@ -4,14 +4,16 @@ const passport = require("passport");
 const bcrypt = require("bcrypt");
 const User = require("../models/User");
 const { generateError, sendError } = require("../helpers/Error");
-const { checkRouteParameters } = require("../helpers/express");
+const { checkRouteParameters } = require("../helpers/Express");
 const Config = require("../../config");
 
 exports.create = async(data, options = {}) => {
+    const { lean } = options;
     if (!Array.isArray(data)) {
         options = null;
     }
-    return await User.create(data, options);
+    const user = await User.create(data, options);
+    return lean ? user.toObject() : user;
 };
 
 exports.find = async(search, projection, options = {}) => await User.find(search, projection, options);
@@ -48,7 +50,7 @@ exports.postUser = async(req, res) => {
     try {
         const { body } = checkRouteParameters(req);
         await this.generateSaltAndHash(body);
-        const newUser = await this.create(body);
+        const newUser = await this.create(body, { lean: true });
         delete newUser.password;
         res.status(200).json(newUser);
     } catch (e) {
@@ -66,16 +68,21 @@ exports.getUsers = async(req, res) => {
 };
 
 exports.login = (req, res) => {
-    passport.authenticate("local", { session: false }, (err, user, info) => {
-        if (err || !user) {
-            return res.status(401).json(generateError("BAD_CREDENTIALS", info));
-        }
-        req.login(user, { session: false }, err => {
-            if (err) {
-                res.status(500).send(err);
+    try {
+        checkRouteParameters(req);
+        passport.authenticate("local", { session: false }, (err, user, info) => {
+            if (err || !user) {
+                return res.status(401).json(generateError("BAD_CREDENTIALS", info));
             }
-            const token = sign({ userId: user._id, exp: Math.floor(Date.now() / 1000) + 3600 * 24 }, Config.app.JWTSecret);
-            return res.status(200).json({ token });
-        });
-    })(req, res);
+            req.login(user, { session: false }, err => {
+                if (err) {
+                    res.status(500).send(err);
+                }
+                const token = sign({ userId: user._id, exp: Math.floor(Date.now() / 1000) + 3600 * 24 }, Config.app.JWTSecret);
+                return res.status(200).json({ token });
+            });
+        })(req, res);
+    } catch (e) {
+        sendError(res, e);
+    }
 };